requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile requirements.in
#
appdirs==1.4.3            # via black
attrs==19.3.0             # via black
backcall==0.1.0           # via ipython
black==19.10b0            # via -r requirements.in
bson==0.5.8               # via djongo
certifi==2020.4.5.1       # via requests
chardet==3.0.4            # via requests
click==7.1.1              # via black
dataclasses==0.6          # via djongo
decorator==4.4.2          # via ipython, traitlets
deprecated==1.2.9         # via pygithub
django-extensions==2.2.9  # via -r requirements.in
django-jsonview==1.3.1    # via -r requirements.in
django==2.2.12            # via -r requirements.in, djongo
djongo==1.3.1             # via -r requirements.in
entrypoints==0.3          # via flake8
flake8==3.7.9             # via -r requirements.in
idna==2.9                 # via requests
ipdb==0.13.2              # via -r requirements.in
ipython-genutils==0.2.0   # via traitlets
ipython==7.13.0           # via -r requirements.in, ipdb
iso8601==0.1.12           # via -r requirements.in
isort==4.3.21             # via -r requirements.in
jedi==0.17.0              # via ipython
mccabe==0.6.1             # via flake8
parso==0.7.0              # via jedi
pathspec==0.8.0           # via black
pexpect==4.8.0            # via ipython
pickleshare==0.7.5        # via ipython
prompt-toolkit==3.0.5     # via ipython
ptyprocess==0.6.0         # via pexpect
pycodestyle==2.5.0        # via flake8
pyflakes==2.1.1           # via flake8
pygithub==1.47            # via -r requirements.in
pygments==2.6.1           # via ipython
pyjwt==1.7.1              # via pygithub
pymongo==3.10.1           # via djongo
python-dateutil==2.8.1    # via bson
pytz==2019.3              # via django
regex==2020.4.4           # via black
requests==2.23.0          # via -r requirements.in, pygithub
six==1.14.0               # via bson, django-extensions, djongo, python-dateutil, traitlets
sqlparse==0.2.4           # via django, djongo
toml==0.10.0              # via black
traitlets==4.3.3          # via ipython
typed-ast==1.4.1          # via black
urllib3==1.25.9           # via requests
wcwidth==0.1.9            # via prompt-toolkit
wrapt==1.12.1             # via deprecated

# The following packages are considered to be unsafe in a requirements file:
# setuptools
